feat: add validation for submodule names
