+++ gix-validate/src/lib.rs
@@ -9,3 +9,6 @@ pub use reference::name as refname;
 ///
 pub mod tag;
 pub use tag::name as tagname;
+
+///
+pub mod submodule;
