--- gix-validate/tests/all.rs
@@ -1,1 +1,0 @@
-mod validate;
