+++ gix-validate/tests/validate.rs
@@ -1,0 +1,3 @@
+mod reference;
+mod submodule;
+mod tag;
