--- gix-validate/tests/validate/mod.rs
@@ -1,2 +1,0 @@
-mod reference;
-mod tagname;
